Fixing linux/windows incompatibility

+++ posit/math_sqrt.hpp
@@ -201,7 +201,7 @@ namespace sw {
 #else
 		template<size_t nbits, size_t es>
 		inline posit<nbits, es> sqrt(const posit<nbits, es>& a) {
-			return posit<nbits, es>(std::sqrtl((long double)a));
+			return posit<nbits, es>(std::sqrt((long double)a));
 		}
 #endif
